load environment variables before loading openai module (and support OPENAI_API_BASE)
OPENAI_API_BASE is a standard environment variable which the openai module will use when it loads. The OPENAI_API_BASE lets you redirect the API to a compatible service. I have such an API in a pull request for oobabooga/text-generation-webui over here: https://github.com/oobabooga/text-generation-webui/pull/1475. With this small change the babyagi code will work unchanged with that.